+++ src/device/cpu/Smp.h
@@ -3,9 +3,8 @@
 
 #include <cstdint>
 
-// Copied from the Linux kernel:
+// Descriptor for either GDT or IDT:
 // https://github.com/torvalds/linux/blob/master/arch/x86/include/asm/desc_defs.h#L112 (visited on 10/02/23)
-// https://github.com/torvalds/linux/blob/master/arch/x86/include/asm/desc.h#L218 (visited on 10/02/23)
 struct Descriptor {
     uint16_t size;
     uint64_t address;
@@ -25,7 +24,7 @@ extern volatile uint32_t boot_ap_entry; // Not written by asm volatile (), so ad
 // Export
 extern "C" [[noreturn]] void smpEntry(uint8_t apicid);
 extern uint32_t** apStacks;
-extern volatile uint8_t runningAPs; // This is used as a bitmap, once an AP is running it sets its corresponding bit to 1
+extern volatile uint64_t runningAPs; // This is used as a bitmap, once an AP is running it sets its corresponding bit to 1
 
 // If any of these two are changed, smp.asm has to be changed too (the %defines at the top)!
 const constexpr uint32_t apStackSize = 0x1000; ///< @brief Size of the stack allocated for each AP.
